libsubprocess/test: setup for alt subprocess write
Problem: libsubprocess iostress tests only support "API" or
"direct" writes in tests.  Currently a simple boolean is used in the
tests to determine which write to use.  But in the near future other
options may be available.

Refactor the iostress tests to use an enum rather than boolean to
indicate how subprocess writes should happen.