python: support addition of sort keys to flux.util.OutputFormat
Problem: There is no unified way to sort items in the output of
commands that use the OutputFormat class in flux.util.

Add support for sort keys directly to the OutputFormat class.

If the text part of the first format element starts with `sort:`, then
the remainder of that text part is taken as a comma-separated list of
sort keys, optionally prefixed by `-` for reverse ordering.  The keys
are split and stored in the OutputFormat instance as a `sort_keys`
attribute, where each element is a tuple of key, reverse pairs.

A new `set_sort_keys()` method is then added to OutputFormat to
override any sort keys set in the original format, e.g. to support
utilities with a `--sort=` command line option.

When callers use the `OutputFormat.print_items()` method, any stored
sort is automatically used. O/w, `OutputFormat.sort_items()` may be
used directly.