flux-overlay: add error response info to trace
Problem: flux overlay trace doesn't show whether response messages
are a success or failure.

Add the errnum value from the message, converted to its short error name,
to the end of the message line.  If there is a human readable error message
in the payload, show it when --full is specified.  Show error responses in
bright red rather than cyan.