libsubprocess: support new on_credit callback
Problem: It would be convenient for libsubprocess users to know when
write buffers have more space available for writing.  However, there
is no mechanism for that.

Support a new on_credit callback that will inform the user that
space is available in the write buffer.  This includes a minor behavior
change to the fbuf write watcher callback.  Adjust tests accordingly.

Fixes #6291