python: store empty string instead of None in FormatSpec attrs
Problem: The OutputFormat.FormatSpec class attempts to convert missing
components of the format spec to an empty string when setting the
components as attributes of the class, but the approach was incorrect,
resulting in attributes with values of None instead of the expected "".

Explicitly convert None values to the empty string when setting
attributes in the class.