python: OutputFormat: only adjust precision for string types
Problem: The OutputFormat `+:` adjusts the precision of a format
spec if the width was adjusted and is now greater than the original
precision.  However, this is only necessary for string types, and in
the case of floating point types will result in more precision than
requested by the original format.

Only make the precision adjustment for string types.