python: handle double-colon in FormatSpec class
Problem: Flux extends the Python formatter to allow a format
spec to follow conversion of a timestamp to a datetime object by
using a double-colon '::'. Because this is an extension, Python's
Formatter.parse() doesn't recognize the format spec, so sometimes part
of the conversion plus the spec will sneak through to the FormatSpec
constructor. The internal regex will then not match and the FormatSpec
will be empty.

Detect this condition by looking for the double-colon. Use the
remainder of the string as the format spec instead.