python: FormatSpec: return empty string not 0 for unset width
Problem: The FormatSpec class returns 0 for spec width when the width
is unspecified, but this causes the OutputFormat.header_format()
method to return `{field:0}` for the header format and this causes
Python 3.6 to raise an exception when formatting the header:

 ValueError: '=' alignment not allowed in string format specifier

Return the empty string in FormatSpec for an unset width so the
corresponding header format will just be `{field}`. This should have
the same result but avoid the Python 3.6 exception.