flux-housekeeping: derive nodelist from ranks in HousekeepingSet
Problem: The internal HousekeepingSet class computes the nodelist
from ranks in the initializer, but this makes it more difficult to
update ranks.

Derive the nodelist from ranks in the property getter. This allows
the ranks associated with a set to be updated without also updating
the nodelist. This is potentially more computationally intensive,
but that is unlikely to matter here.