flux-housekeeping: combine nodes/ranks for otherwise like lines
Problem: It would be useful to get the set of all ranks/nodes
in housekeeping with `flux housekeeping -no {nodelist}`, but this
currently does not work because the output for each job is placed on
a separate line even if jobid is not displayed.

Use the Deduplicator in `flux housekeeping list` to enable combination
of otherwise duplicate lines.