python: JournalConsumer: detect poll() called without start()
Problem: If a JournalConsumer object's poll() method is called before
start(), then self.rpc will be None and Python will raise an exception:

  AttributeError: 'NoneType' has no attribute 'wait_for'

or similar. It would be better to raise a more helpful error.

Detect if `poll()` is called before `start()` and raise a RuntimError
with helpful error text.