python: JournalConsumer: allow set_callback() before start()
Problem: The JournalConsumer set_callback() method calls
self.rpc.then(), which implies that start() must be called before
set_callback(). But this is too restrictive and unnecessary.

Call then() in start() if a callback is set. This allows a user to
set the callback either before or after start().

Also, return self from set_callback(). This is sometimes convenient
to chain class like

 consumer = JournalConsumer(handle).set_callback(callback)