broker: detect mismatched bootstrap.hosts config
Problem: if multiple hosts think they are the same rank, each
will cause the other to be disconnected, flipping back and forth,
logging a "reconnected after crash" error that is misleading.

Send the hostname with to the "overlay.hello" request.  If the
rank and hostname presented in the hello request do not match
the expected configuration, log a useful message and disconnect
the misconfigured host.

The broker with the correctly configured host is not disturbed.

Update overlay unit test.

Fixes #6389