testsuite: use fake hosts in t4182-resource-rerank
Problem: t4182-resource-rerank.sh will fail when the broker checks
that clients are connecting with the expected hostname:rank mapping.

The test is only setting the hostlist broker attribute, but this attribute
will now be used to check the hostname:rank mapping in overlay.hello,
thus the test brokers are rejected and the test falls apart.

Change the test so that it uses flux start --test-hosts, which sets
FLUX_FAKE_HOSTNAME for each rank AND the hostlist broker attribute.
Now the broker presents the fake hostname to overlay.hello.

In the sub-tests that run within the test instance, simulate --test-hosts
for a job by providing a broker wrapper script that accepts the fake hostlist
and sets the hostlist attribute and FLUX_FAKE_HOSTNAME to the broker's index
into the hostlist according to FLUX_TASK_RANK.

Disable resource verification in the subtests so that rlist_verify() does
not fail when libhwloc finds the fake hostname is not the system one.