broker: add optional cleanup timeout on signal
Problem: 'systemctl stop flux' on rank 0 may hang waiting for the
queue to become idle, and be killed before it can write its KVS
checkpoint if the systemd 90s TimeoutStopSec is exceeded.

Optionally impose a timeout on cleanup if the broker is shutting down
due on receipt of a terminating signal.  Note that the timeout does
NOT apply when flux-shutdown(1) is used to shut down the instance.

broker.cleanup-timeout must be explicitly configured to enable this
(default value is "none").