python: fix queue-specific override of defaults in frobnicator plugin
Problem: The `defaults` job frobnicator plugin fails to override
global defaults with queue defaults because the global default is
applied to the jobspec before the queue defaults are examined, and
since a value is already set in jobspec, the queue default is ignored.

Instead of applying global defaults before examining queue defaults,
update the `queue_defaults()` method to take a defaults dict and
update it with queue specific defaults. Then the combined defaults
are applied in one pass, avoiding the issue above.

Fixes #6400