perilog: ensure job doesn't start when prolog cancel times out
Problem: When the a prolog is canceled, but the cancel times out, no
job exception is raised and the job proceeds to start. No only this,
but the epilog can end up being run twice since there's an assumption
in the code that there will be no finish event.

Add a new flag to the perilog_proc struct to capture when a canceled
prolog times out. Emit a specific exception in emit_finish_event()
when this flag is set and force the prolog exit status to 1.