docs: rfc flux-config-bootstrap diagram
Problem: It is really easy to miss the subtle comma (instead
of a dash) for one of the bootstrap directives, resulting in
a lot of confusion trying to figure out how the tree makes
sense.
Solution: Provide a visual diagram of the tree and some
additional explanation to make it explicitly clear.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>