job-manager/history: optimize list insertion
Problem: Internal job history lists are sorted with larger
t_submit values at the front of the list.  When new jobs
are inserted (i.e. jobs with bigger t_submit values), they
are inserted with a search starting at the back of the list.
This is the opposite of what should be done and leads to a
slowdown in job throughput over time.

Insert into the job history list starting at the front.

Fixes #6419