libutil: add timestamp_tzoffset()
Problem: It would be useful to optionally append the TZ offset to
ISO 8601 timestamps in Flux, but no function exists to do that.

Add a new timestamp_tzoffset() function that takes a struct tm and
returns the current timezone TZ offset of the form [+-]HH:MM. Note
that this function is required in order to add the colon (:), which
aids in readability, because strftime(3) "%z" does not include a colon.
(The colon is optional in ISO 8601).

timestamp_tzoffset() also returns "Z" instead of "+00:00" for UTC for
backwards compatibility with existing timestamp formats.