libflux: update config tables recursively in flux_conf_parse(3)
Problem: When loading configuration via multiple TOML or
JSON files from a directory, tables from files that update the
configuration object replace existing tables in the config instaed
of updating them. This is because the conf object is updated with
json_object_update(3), which is not recursive.

Replace use of json_object_update() with json_object_update_recursive()
so that tables are updated recursively instead of replaced.

Fixes #6396