libflux/test: test that flux conf tables update recursively
Problem: No tests in the flux_conf_t unit tests ensure that a table
present in a conf file doesn't overwrite an existing table.

Expand the unit tests in libflux/test/conf.c to add a fourth config
file that updates a table. Ensure that the value is added without
overwriting the existing key.