python: indicate held jobs in JobInfo.contextual_info field
Problem: Held jobs (those with priority=0) are not denoted in default
flux-jobs(1) output, but this would be useful for admins and users
to quickly determine why these jobs are not being scheduled.

Update the `contextual_info` property of the `JobInfo` class (which
supplies the content for the INFO column) to return "held" for jobs
with urgency=0 and "priority-hold" for jobs with priority=0.

Fixes #6426