libflux: move reactor ops out of public API
Problem: adding operation to 'struct watcher_ops' breaks ABI.

Drop that struct and associated functions from the public API.
There are no known watcher implementations outside of flux-core,
and these functions are undocumented.

We can re-publish this part of the API when there is demand, and
take more care to make it it resilient to change (e.g. adding padding
to the struct) at that time.

Update one user to include the private header.

Fixes #6433