python: support `wait` parameter to jobid URI resolver
Problem: It is inconvenient that a user has to ensure a job's
memo event containing the subinstance URI has been posted before
they can call `flux uri` or other URI resolver.

Support a `wait` parameter for the jobid resolver which, when used,
causes the resolver to wait for the `memo` event containing the
URI before returning. This is only done if the job is not in
the INACTIVE or CLEANUP states.