flux-job: attach status doesn't handle queue updates
Problem: The flux-job attach status line may indicate that a job is
waiting for resources due to a stopped queue, even after the job's
queue has been updated to a started queue.

Always refresh the queue name after a `jobspec-update` event in case
the queue has changed. This will force the status message to update
as well once the new queue and its status is available.