job-manager: support partial-ok in hello request
Problem: RFC 27 allows the scheduler to send a partial-ok flag
in the hello request, and then receive partially allocated jobs
in hello responses.

If the hello request includes this flag, pass it on to housekeeping.
For each partially released housekeeping job, include the 'free'
idset in the response per RFC 27.