flux-job: attach: improve output of exception with status line
Problem: When a job exception is emiitted by `flux job attach` and
the statusline is active, the job exception may not overwrite the
entire line and is garbled with the status message.

If the statusline is active, emit an ansi escape sequence to clear
the current line before printing the job exception message.