libschedutil: support hello 'free' key
Problem: when processing hello responses, all schedulers now need
to process R - free for partial releases.

As a convenience, change the libschedutil hello callback to subtract
the free idset from the R it fetched from the KVS.

Note that the scheduling key, if present, remains the full object
which is opaque to flux-core.