kvs-watch: avoid error response on failed rpc
Problem: In the function handle_lookup_response(), if a call to
flux_respond_pack() fails, an additional call to flux_respond_error()
can be made in the "goto error".  This is not common usage in flux-core.
If flux_respond_pack() fails, we simply log an error and avoid a second
attempt to send a response.

In handle_lookup_response() do not automatically call flux_respond_error()
on an error.  In called function, call flux_respond_error() if necessary.