doc: improve `--include` documentation in flux-resource(1)
Problem: The flux-resource(1) documentation of the `-i, --include`
option doesn't include the word "node", which made is less likely
the option would pop up for a search on how to filter nodes in the
output of various subcommands.

Include "node" in the description of the `-i, --include` option. Make
the text for all instances of the `--include` option the same.