resource: only read resource.scheduling config on rank 0
Problem: The Rv1 scheduling key as JGF can be large and time-consuming
to generate and distribute, but currently it is required on all
nodes. Since the key is really only used by the scheduler on rank 0,
it would be ideal if resource.scheduling was only read on rank 0.

Only process the resource.scheduling config key on rank 0.

Fixes #6480