testsuite: add test for job shell handling of SIGPIPE
Problem: There's no test to ensure the job shell is not killed by
SIGPIPE when writing data to a task that has closed stdin.

Attempt to add a test for this issue. Add a note that the test is
inherently racy, and it is just a best effort at this time.