sdbus: fix watcher is_active() callback
Problem: the internal sdbus watcher is_active() callback calls
ev_is_active() on a flux_watcher_t, which is accepted because
ev_is_active() is a macro that casts its pointer argument.

Although flux_watcher_is_active() is not currently used in the
sdbus module (the only user of the watcher), fix the code in
case it ever is.