job-info: stream events when job is inactive
Problem: If a job is inactive, all data in an eventlog will be retrieved
from the KVS in a single lookup.  This is because we know the data should never
change after the job is inactive.

If this data is very large, this lookup can be slow.  In some cases, the
use of something like `flux job attach` can have the appearance of a hang
because the standard output response is taking so long to lookup and return.

Solution:

When a job is inactive and the user wants to watch a job eventlog,
do not retrieve all of the data from the KVS in a single lookup.  Instead,
use the FLUX_KVS_STREAM flag to retrieve the data in smaller chunks.  This data
will be internally read and parsed no differently than when the job is active.

Fixes #6516