job-info: stream events when job is inactive
Problem: If a job is inactive, all data in an eventlog will be retrieved
as a single response during an eventlog watch.  This is because we know for
a fact that the data should never change after the job is inactive.

If this data, such as job standard output, is very large, this lookup
can be very slow.  In some cases, the use of something like `flux job attach`
can have the appearance of a hang because the standard output response is
taking so long to lookup and return.

Solution:

When a job is inactive and the user wants to watch a job eventlog,
do not retrieve all of the data.  Instead, retrieve the data via an internal
eventlog watch, but have the eventlog watch use the new FLUX_KVS_STREAM
flag.

Fixes #6516