python: cli: fix handling of newline in `--add-file`
Problem: Documentation for the `--add-file=NAME=DATA` option indicates
that DATA will be treated as a path unless a newline appears in DATA.
However, the obvious doesn't work:

 --add-file=test='this is a test file\n'

because what Python actually gets is an escaped newline '\\n', so
test for '\n' in the string fail.

Convert any escaped newlines to actual newlines in
handle_add_file_arg() so that subsequent detection of newline works.