python: cli: allow mustache templates in --env* options
Problem: It would be useful to expand mustache templates in environment
variable values, but this is currently not supported in the submission
cli commands.

When processing the `--env*` set of options, detect any provided
environment variable values that appear to contain a mustache
template. Return these in a separate dict from the main environ
dictionary and add them to `attributes.system.shel.optionsenv-expand`
in jobspec.  The job shell will then expand these variables before
running the job.