shell: support expansion of select env vars as mustache templates
Problem: It would be useful to allow expansion of mustache templates
in envionment variables in the job shell. However, expanding templates
in the entire jobspec environment dictionary would be inefficient,
since values with templates would be by far the uncommon case. Also,
environment variables may legitimately contain values that look like
mustache templates but should not be expanded.

Add a new shell builtin plugin for expanding select environment
variables in a `env-expand` shell options dictionary. This allows
the jobspec to opt-in to mustache expansion for a select set of
environment variables, solving the issues noted above.