python: allow perms=None to get default in `Jobspec.add_file()`
Problem: If perms is passed as None in Jobspec.add_file(), then
the default of 0600 is overridden. It would be better to use the
default in this case.

Set the default value of perms to None in the function arguments,
and if perms is None use the default of 0o600.