kvs-watch: support FLUX_KVS_STREAM flag
Problem: Responses to KVS lookups must be contained within one message,
no matter how large the value is.  Such a response can be slow and give the
appearance that something is hanging.  For example, a large amount of standard
output stored in the KVS would not return immediately.

Support a new FLUX_KVS_STREAM flag.  It will use the mechanisms built
into the kvs-watch module and the FLUX_KVS_WATCH_APPEND flag to stream
each content blob of large values to the user.  Once all the content is
streamed, it will return ENODATA to indicate the stream is complete.