kvs-watch: support FLUX_KVS_STREAM flag
Problem: It would be convenient if there was a way to stream large
data from the KVS by their individual content blobs, rather than getting
a very large piece of content in one response.  Such a response can be
slow and give the appearance that something is hanging.  For example, this
ability would be useful for a large amount of standard output stored in the
KVS.

Support a new FLUX_KVS_STREAM flag.  It will use the mechanisms built
into the kvs-watch module and the FLUX_KVS_WATCH_APPEND flag to stream
each content blob to the user.  Once all the content is streamed, it will
return ENODATA to indicate the stream is complete.