env-expand: allow per-task environment variables if supported
Problem: Environment variables in the env-expand jobspec options
dictionary cannot be expanded to different values per task because
the env-expand plugin only runs at the shell.init callback.

Add a task_expand_env() function in the env-expand plugin that runs
at task.init and thus is able to set per-task environment variables.

To avoid processing the same mustache templates multiple times,
keep the original env_expand() function to expand non task-specific
templates.  When a template is detected to be fully expanded, it is
removed from the env-expand object so that it is not re-processed
per task.