job-exec: improve drain message for unkillable procs
Problem: The drain message issued by the job-exec module when it
gives up trying to kill user processes in the job is not clear.

Note that the unkillable job procesess were user processes to avoid
confusion that epilog or housekeeping processes were at fault.

Update one test that attempts to match the old error message.

Fixes #6495