shell: support node-specific tags in mustache template
Problem: The shell doesn't support expansion of any node-specific
tags in mustache templates.

Add support for expansion of node-specific data via a set of node.<key>
mustache tags. The <key> is pulled directly from the shell "rank info"
object, so any item present in that object can be expanded, including
{{node.id}} for the nodeid (job local node rank), {{node.name}} for
the hostname, {{node.taskids}} for the list of tasks assigned to the
node, etc.