shell: fix signal plugin to work with Fluxion scheduler
Problem: The shell signal plugin doesn't work with Fluxion because
it expects a float value for expiration in R and Fluxion encodes
an integer.

Update the format specifier in the signal plugin's use of json_unpack
on R to use `s:F` instead of `s:f` so that the plugin works with
Fluxion.

Fixes #6521