shell: do not log error on uknown mustache template
Problem: Callers to flux_shell_mustache_render() may want to know
if a tag or tags is not supported, but not necessarily have an
error logged to the job.

Return an error with errno=ENOENT instead of logging an error when
a tag is not supported and it is practical to do so.