shell: allow custom task, shell_rank passed to mustache_render()
Problem: The shell mustache render functionality only allows rendering
of templates for the current task and shell rank, but it may be useful
to use this functionality for an arbitrary task/node.

Add a new `struct mustache_arg` which is passed to the renderer.
For now, fill it in with the shell, current task, and current rank.