shell: add rank/task specific mustache render functions to api
Problem: There is no way for shell plugins to check if a mustache
template would render to a different result on a different task or
rank, because flux_shell_mustache_render(3) only works for the current
rank and task.

Add new functions to render mustache templates for any shell rank
or local task.