python: fix precision OutputFormat expandable width
Problem: The `+:` sentinel which expands width of fields doesn't adjust
precision as well, which ends up truncating strings in a field that
has been expanded to fit the largest value.

The set of matching types needs to be expanded from ("s", None)
to include the empty string to handle the case where a presentation
type is not specified, so string is assumed.

Add "" to the set of types for which to adjust precision.

Fixes #6530