job-exec: don't drop stdout before shell barrier
Problem: Early output from the job shell may be dropped if it comes
before the shell barrier. This is because the shell barrier code in
the job-exec module throws away data that doesn't exactly match the
string "enter\n".

If the current line doesn't match "enter\n", log the line to the
job's exec eventlog so that `flux job attach` etc can dump the output.