testsuite: add test for `flux hostlist avail` with down nodes
Problem: The tests of the flux-hostlist command do not check that
the hostlist returned from `flux hostlist avail` does not include
down/offline nodes.

Convert the t2814-hostlist-cmd.t test to use the 'system'
test_under_flux personality. Remove a rank and ensure that rank is
not reported in `flux hostlist avail`.