testsuite: avoid `shell.log` in invalid-args shell tests
Problem: The shell plugin `invalid-args.so` subscribes to all shell.*
plugin topics and runs a set of tests with invalid arguments to shell
API functions. However, use of the `shell.log` topic here has been
shown to cause a deadlock due to it being recursively called from
`task.exec`: the stderr buffer becomes full and the task gets blocked,
which prevents it from entering the shell barrier.

Skip testing for the `shell.lost` callback. It doesn't make sense
to test the shell API invalid args from this callback topic anyway,
since these functions are not likely to be used there.