shell: pty: pack shell.output args in RFC 24 data event format
Problem: The shell.output plugin topic args (currently only used
in the job shell pty plugin) is mostly in RFC 24 data event format,
except that the rank is encoded as an integer instead of a string. This
means the plugin args have to be unpacked and repacked into a context
suitable for use with iodecode().

Pack shell.output args with a string rank so the object can be passed
directly to iodecode(). This will remove unnecessary decode and encode
plus memory allocation in the output plugin.