libsubprocess: avoid using child watchers
Problem: child watchers are only used by libsubprocess, and probably
do not need to be in the public API since libsubprocess is provided
for managing child processes.  In addition, we are considering porting
Flux to libuv and libuv does not offer similar functionality.

Register a signal watcher for SIGCHLD within libsubprocess that
persists as long as there are subprocesses to monitor, and calls
waitpid(2) to consume all subprocess state changes.  Add a hash by pid
and allow subprocess objects to register a callback to receive these
changes for a given pid.

Have all subprocess users create reactors without FLUX_REACTOR_SIGCHLD,
as the default ev_loop registers a SIGCHLD watcher that conflicts with
this one.

Add EVFLAG_SIGNALFD to the non-default loop, as that flag was used with
reactors created with FLUX_REACTOR_SIGCHLD, and appears to be be required
to avoid sharness tests hanging randomly.