resource: support resource.reserve configuration key
Problem: There is no way to reserve a set of cores so they cannot
be used by the scheduler of a Flux instance.

Add support for a new config key `resource.reserve`, which takes a
string of the form `cores[@ranks]` where `cores` and the optional
`ranks` are idsets specifying the cores to reserve and the ranks
on which to reserve them. If ranks is not specified, then the spec
applies to all ranks.

The reserved set of cores are subtracted from the resource set before
it is handed off to the scheduler. The reserved resource set is also
removed from the status response used by `flux resource list`.