libflux: add libuv support to reactor
Problem: the flux reactor lacks an interface for embedding
in other reactors.

libev doesn't offer access to a file descriptor that becomes
ready when events are pending.  However, libuv provides
uv_backend_fd().  Therefore, let's begin a port of libuv
to Flux.

This commit creates an alternate libflux reactor implementation
based on libuv that is selected when flux-core is configured
--with-libuv.  Reactor watchers are dealt with in a separate
commit.