libsubprocess/test: Fix racy fd count in test
Problem: A file descriptor counting test accesses /proc/PID/fd to
count file descriptors.  The `ls` used in the test may lead to an
additional file descriptor that is not accounted for.

Increase the expected number of file descriptors in the test to account
for the one used with `ls`.

Fixes #6551