libflux: avoid truncation of log messages
Problem: Flux log messages may be truncated because all messages
are formated into an internal fixed size buffer of FLUX_MAX_LOGBUF
(2048) bytes. On large clusters, this can result in regular loss of
important information in the logs.

Attempt to allocate a large enough buffer for the full log message
on the heap when truncation is detected in flux_vlog(3). Do this
for both the case where the log destination is stderr (h == NULL),
or the message is being forwarded to a handler or via an RPC.