libflux: test: test a large message in flog unit tests
Problem: There is no test of logging a large message to stderr in
the libflux log unit tests.

Add a test that logs a message > FLUX_MAX_LOGBUF.

Note: Since the log message goes to stderr, i.e. the output of the
unit test itself, we can't ensure the message was not truncated. However,
this is still useful to ensure the test does not die, and that no leaks
are present (i.e. by running the test under valgrind)