cmd: add `--root` option to `flux(1)`
Problem: While the `flux(1)` command supports operating on parent
URIs by specifying one or more `--parent` options on the command line,
there is no way to easily specify the top-level or root parent.

Add a `--root` option which acts if `instance-level` number of
`--parent` options were supplied to the command.