python: add optional remote_hostname arg to JobURI()
Problem: When an instance of the flux.job.uri.JobURI class is
initialized with a local URI and it is converted to a remote URI,
the local hostname is always used. However, sometimes this may
not be appropriate.

Add an optional `remote_hostname` arg to the JobURI initializer.
If set, use that hostname instead of the local hostname.