python: support "relative" URIs in the jobid URI scheme
Problem: There is currently no way to easily obtain the URI of a
parent Flux instance, the root instance, etc.

The `jobid` URI scheme already supports jobid "paths" like `id1/id2`
to obtain the URI of a job `id2` running in instance `id1`. Extend
this support with familiar syntax for "relative" jobid paths such as
`/` for the top-level (or root) instance, or `..` for the local parent,
or `/jobid` for a jobid running within in the root instance (no matter
the current instance level at which the query is made)

While not as useful, `jobid/..` is also supported, where the hostname
part of the URI for jobid is then inherited by the URI returned from
its parent (by default all relative URIs are resolved as local URIs)