broker: set FLUX_ENCLOSING_ID when jobid attribute is set
Problem: It is inconvenient to get the jobid of the closest enclosing
instance (e.g. the ID of a batch or alloc job in the parent instance)
because the broker drops FLUX_JOB_ID in favor of a jobid attribute.
However, a case has been made that an environment variable would be
more convenient since it can be accessed without use of system(3)
or the Flux API.

Introduce FLUX_ENCLOSING_ID, which is set by the broker whenever the
jobid attribute is set (i.e. when the broker is started as a job in
a Flux instance). This will be available in the initial program as
well as being inherited by jobs run within the instance.

Add the variable to the env_blocklist so that it is unset when the
current instance is not a job.

Fixes #6474