broker: only set parent-uri if instance is a job
Problem: The broker uses presence of FLUX_JOB_ID in the environment
to determine if it needs to set the parent-uri attribute, but this
ends up setting parent-uri in cases where the instance is not a job
such as `flux run flux start --test-size=1`.

Use the `jobid` broker attribute instead of FLUX_JOB_ID. This attribute
is only set by boot_pmi() if the instance was really started as a
job in another Flux instance.