t: move some tests towards end of tests
Problem: The testing of stats clearing in the KVS is done
a little bit before the end of a set of tests.  This can be
inconvenient for future tests that do not wish stats to be cleared.

Move the clearing of stats tests towards the end of tests in
t1001-kvs-internals.t.