testsuite: fix flaky test in t0019-tbon-config.t
Problem: The test `tbon.endpoint cannot be set` in t0019-tbon-config.t
fails occasionally in ci because the `flux start` command reports it
is terminated by SIGPIPE rather than SIGKILL/SIGTERM.

The test is just meant to detect an expected failure of the command, it
does not matter how the command fails. Just use `!` to detect failure.
Add a note to explain why `test_must_fail` or similar is not used here.

Fixes #6336