dependency-after: ensure only jobs that ran satisfy dependencies
Problem: The dependency-after plugin allows jobs that were canceled
before the RUN state to satisfy an `afternotok` dependency.  This
causes unexpected behavior when `afternotok` chaining is used to
keep a job running until success: Once the job is successful, this
raises an exception on the dependent job, which fails and releases
the depedency on the next job in the chain, which unexpectedly runs.

The principle of least surprise probably dictates that `after*`
dependencies can only ever be satified for jobs that actually ran,
and therefore jobs canceled before an `alloc` event should raise an
exception on *any* dependent job.

Update the dependency-after plugin so that depencies are never satified
for jobs that did not enter the RUN state.

Fixes #6555